Update to 1.5.1
- Update Subject model since `oneOf` does not seem to be working correctly, causing validation errors when parsing the Measurement API response
