+++ .openapi-generator/FILES
@@ -49,8 +49,6 @@ docs/ResultReport.md
 docs/Results.md
 docs/Specimen.md
 docs/Subject.md
-docs/SubjectOneOf.md
-docs/SubjectOneOf1.md
 docs/V2Device.md
 docs/V2Dosage.md
 docs/V2Medication.md
@@ -110,8 +108,6 @@ lib/primary_connect_client/models/result_report.rb
 lib/primary_connect_client/models/results.rb
 lib/primary_connect_client/models/specimen.rb
 lib/primary_connect_client/models/subject.rb
-lib/primary_connect_client/models/subject_one_of.rb
-lib/primary_connect_client/models/subject_one_of1.rb
 lib/primary_connect_client/models/v2_device.rb
 lib/primary_connect_client/models/v2_dosage.rb
 lib/primary_connect_client/models/v2_medication.rb
@@ -169,8 +165,6 @@ spec/models/result_report_spec.rb
 spec/models/result_spec.rb
 spec/models/results_spec.rb
 spec/models/specimen_spec.rb
-spec/models/subject_one_of1_spec.rb
-spec/models/subject_one_of_spec.rb
 spec/models/subject_spec.rb
 spec/models/v2_device_spec.rb
 spec/models/v2_dosage_spec.rb
