+++ README.md
@@ -7,7 +7,7 @@ Client Library to interface with Primary Connect
 This SDK is automatically generated by the [OpenAPI Generator](https://openapi-generator.tech) project:
 
 - API version: v1
-- Package version: 1.5.0
+- Package version: 1.5.1
 - Build package: org.openapitools.codegen.languages.RubyClientCodegen
 
 ## Installation
@@ -23,16 +23,16 @@ gem build primary_connect_client.gemspec
 Then either install the gem locally:
 
 ```shell
-gem install ./primary_connect_client-1.5.0.gem
+gem install ./primary_connect_client-1.5.1.gem
 ```
 
-(for development, run `gem install --dev ./primary_connect_client-1.5.0.gem` to install the development dependencies)
+(for development, run `gem install --dev ./primary_connect_client-1.5.1.gem` to install the development dependencies)
 
 or publish the gem to a gem hosting service, e.g. [RubyGems](https://rubygems.org/).
 
 Finally add this to the Gemfile:
 
-    gem 'primary_connect_client', '~> 1.5.0'
+    gem 'primary_connect_client', '~> 1.5.1'
 
 ### Install from Git
 
@@ -142,8 +142,6 @@ Class | Method | HTTP request | Description
  - [PrimaryConnectClient::Results](docs/Results.md)
  - [PrimaryConnectClient::Specimen](docs/Specimen.md)
  - [PrimaryConnectClient::Subject](docs/Subject.md)
- - [PrimaryConnectClient::SubjectOneOf](docs/SubjectOneOf.md)
- - [PrimaryConnectClient::SubjectOneOf1](docs/SubjectOneOf1.md)
  - [PrimaryConnectClient::V2Device](docs/V2Device.md)
  - [PrimaryConnectClient::V2Dosage](docs/V2Dosage.md)
  - [PrimaryConnectClient::V2Medication](docs/V2Medication.md)
