+++ lib/primary_connect_client.rb
@@ -59,8 +59,6 @@ require 'primary_connect_client/models/result_report'
 require 'primary_connect_client/models/results'
 require 'primary_connect_client/models/specimen'
 require 'primary_connect_client/models/subject'
-require 'primary_connect_client/models/subject_one_of'
-require 'primary_connect_client/models/subject_one_of1'
 require 'primary_connect_client/models/v2_device'
 require 'primary_connect_client/models/v2_dosage'
 require 'primary_connect_client/models/v2_medication'
