+++ lib/primary_connect_client/models/measurement.rb
@@ -48,7 +48,6 @@ module PrimaryConnectClient
     # List of attributes with nullable: true
     def self.openapi_nullable
       Set.new([
-        :'subject',
       ])
     end
