+++ lib/primary_connect_client/models/order.rb
@@ -55,7 +55,6 @@ module PrimaryConnectClient
     # List of attributes with nullable: true
     def self.openapi_nullable
       Set.new([
-        :'subject',
       ])
     end
