--- lib/primary_connect_client/models/patient_demographics.rb
@@ -1,466 +1,0 @@
-=begin
-#Primary Connect API V1
-
-#Client Library to interface with Primary Connect
-
-The version of the OpenAPI document: v1
-
-Generated by: https://openapi-generator.tech
-OpenAPI Generator version: 5.3.0
-
-=end
-
-require 'date'
-require 'time'
-
-module PrimaryConnectClient
-  class PatientDemographics
-    attr_accessor :name
-
-    attr_accessor :dob
-
-    attr_accessor :ssn
-
-    attr_accessor :sex
-
-    attr_accessor :gender
-
-    attr_accessor :sexual_orientation
-
-    attr_accessor :races
-
-    attr_accessor :race_details
-
-    attr_accessor :ethnicity
-
-    attr_accessor :ethnicity_detail
-
-    attr_accessor :marital_status
-
-    attr_accessor :is_deceased
-
-    # ISO 8601
-    attr_accessor :death_date_time
-
-    attr_accessor :phone_numbers
-
-    attr_accessor :email_addresses
-
-    attr_accessor :language_code
-
-    attr_accessor :citizenships
-
-    attr_accessor :address
-
-    class EnumAttributeValidator
-      attr_reader :datatype
-      attr_reader :allowable_values
-
-      def initialize(datatype, allowable_values)
-        @allowable_values = allowable_values.map do |value|
-          case datatype.to_s
-          when /Integer/i
-            value.to_i
-          when /Float/i
-            value.to_f
-          else
-            value
-          end
-        end
-      end
-
-      def valid?(value)
-        !value || allowable_values.include?(value)
-      end
-    end
-
-    # Attribute mapping from ruby-style variable name to JSON key.
-    def self.attribute_map
-      {
-        :'name' => :'name',
-        :'dob' => :'dob',
-        :'ssn' => :'ssn',
-        :'sex' => :'sex',
-        :'gender' => :'gender',
-        :'sexual_orientation' => :'sexual_orientation',
-        :'races' => :'races',
-        :'race_details' => :'race_details',
-        :'ethnicity' => :'ethnicity',
-        :'ethnicity_detail' => :'ethnicity_detail',
-        :'marital_status' => :'marital_status',
-        :'is_deceased' => :'is_deceased',
-        :'death_date_time' => :'death_date_time',
-        :'phone_numbers' => :'phone_numbers',
-        :'email_addresses' => :'email_addresses',
-        :'language_code' => :'language_code',
-        :'citizenships' => :'citizenships',
-        :'address' => :'address'
-      }
-    end
-
-    # Returns all the JSON keys this model knows about
-    def self.acceptable_attributes
-      attribute_map.values
-    end
-
-    # Attribute type mapping.
-    def self.openapi_types
-      {
-        :'name' => :'Name',
-        :'dob' => :'String',
-        :'ssn' => :'String',
-        :'sex' => :'String',
-        :'gender' => :'String',
-        :'sexual_orientation' => :'String',
-        :'races' => :'Array<String>',
-        :'race_details' => :'Array<String>',
-        :'ethnicity' => :'String',
-        :'ethnicity_detail' => :'Array<String>',
-        :'marital_status' => :'String',
-        :'is_deceased' => :'Boolean',
-        :'death_date_time' => :'Time',
-        :'phone_numbers' => :'Array<PhoneNumber>',
-        :'email_addresses' => :'Array<String>',
-        :'language_code' => :'String',
-        :'citizenships' => :'Array<String>',
-        :'address' => :'Address'
-      }
-    end
-
-    # List of attributes with nullable: true
-    def self.openapi_nullable
-      Set.new([
-      ])
-    end
-
-    # Initializes the object
-    # @param [Hash] attributes Model attributes in the form of hash
-    def initialize(attributes = {})
-      if (!attributes.is_a?(Hash))
-        fail ArgumentError, "The input argument (attributes) must be a hash in `PrimaryConnectClient::PatientDemographics` initialize method"
-      end
-
-      # check to see if the attribute exists and convert string to symbol for hash key
-      attributes = attributes.each_with_object({}) { |(k, v), h|
-        if (!self.class.attribute_map.key?(k.to_sym))
-          fail ArgumentError, "`#{k}` is not a valid attribute in `PrimaryConnectClient::PatientDemographics`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
-        end
-        h[k.to_sym] = v
-      }
-
-      if attributes.key?(:'name')
-        self.name = attributes[:'name']
-      end
-
-      if attributes.key?(:'dob')
-        self.dob = attributes[:'dob']
-      end
-
-      if attributes.key?(:'ssn')
-        self.ssn = attributes[:'ssn']
-      end
-
-      if attributes.key?(:'sex')
-        self.sex = attributes[:'sex']
-      end
-
-      if attributes.key?(:'gender')
-        self.gender = attributes[:'gender']
-      end
-
-      if attributes.key?(:'sexual_orientation')
-        self.sexual_orientation = attributes[:'sexual_orientation']
-      end
-
-      if attributes.key?(:'races')
-        if (value = attributes[:'races']).is_a?(Array)
-          self.races = value
-        end
-      end
-
-      if attributes.key?(:'race_details')
-        if (value = attributes[:'race_details']).is_a?(Array)
-          self.race_details = value
-        end
-      end
-
-      if attributes.key?(:'ethnicity')
-        self.ethnicity = attributes[:'ethnicity']
-      end
-
-      if attributes.key?(:'ethnicity_detail')
-        if (value = attributes[:'ethnicity_detail']).is_a?(Array)
-          self.ethnicity_detail = value
-        end
-      end
-
-      if attributes.key?(:'marital_status')
-        self.marital_status = attributes[:'marital_status']
-      end
-
-      if attributes.key?(:'is_deceased')
-        self.is_deceased = attributes[:'is_deceased']
-      end
-
-      if attributes.key?(:'death_date_time')
-        self.death_date_time = attributes[:'death_date_time']
-      end
-
-      if attributes.key?(:'phone_numbers')
-        if (value = attributes[:'phone_numbers']).is_a?(Array)
-          self.phone_numbers = value
-        end
-      end
-
-      if attributes.key?(:'email_addresses')
-        if (value = attributes[:'email_addresses']).is_a?(Array)
-          self.email_addresses = value
-        end
-      end
-
-      if attributes.key?(:'language_code')
-        self.language_code = attributes[:'language_code']
-      end
-
-      if attributes.key?(:'citizenships')
-        if (value = attributes[:'citizenships']).is_a?(Array)
-          self.citizenships = value
-        end
-      end
-
-      if attributes.key?(:'address')
-        self.address = attributes[:'address']
-      end
-    end
-
-    # Show invalid properties with the reasons. Usually used together with valid?
-    # @return Array for valid properties with the reasons
-    def list_invalid_properties
-      invalid_properties = Array.new
-      invalid_properties
-    end
-
-    # Check to see if the all the properties in the model are valid
-    # @return true if the model is valid
-    def valid?
-      sex_validator = EnumAttributeValidator.new('String', ["SEX_UNKNOWN", "SEX_FEMALE", "SEX_MALE", "SEX_OTHER", "SEX_NON_BINARY", "SEX_PREFER_NOT_TO_DISCLOSE"])
-      return false unless sex_validator.valid?(@sex)
-      gender_validator = EnumAttributeValidator.new('String', ["GENDER_UNKNOWN", "GENDER_FEMALE", "GENDER_MALE", "GENDER_OTHER", "GENDER_NON_BINARY", "GENDER_TRANSGENDER_MALE", "GENDER_TRANSGENDER_FEMALE", "GENDER_PREFER_NOT_TO_DISCLOSE"])
-      return false unless gender_validator.valid?(@gender)
-      sexual_orientation_validator = EnumAttributeValidator.new('String', ["ORIENTATION_UNKNOWN", "ORIENTATION_HOMOSEXUAL", "ORIENTATION_HETEROSEXUAL", "ORIENTATION_BISEXUAL", "ORIENTATION_QUESTIONING", "ORIENTATION_OTHER", "ORIENTATION_PREFER_NOT_TO_DISCLOSE"])
-      return false unless sexual_orientation_validator.valid?(@sexual_orientation)
-      ethnicity_validator = EnumAttributeValidator.new('String', ["ETHNICITY_UNKNOWN", "ETHNICITY_HISPANIC", "ETHNICITY_NON_HISPANIC", "ETHNICITY_PREFER_NOT_TO_DISCLOSE"])
-      return false unless ethnicity_validator.valid?(@ethnicity)
-      marital_status_validator = EnumAttributeValidator.new('String', ["MARITAL_STATUS_UNKNOWN", "MARITAL_STATUS_SEPARATED", "MARITAL_STATUS_UNMARRIED", "MARITAL_STATUS_COMMON_LAW", "MARITAL_STATUS_DIVORCED", "MARITAL_STATUS_LEGALLY_SEPARATED", "MARITAL_STATUS_LIVING_TOGETHER", "MARITAL_STATUS_INTERLOCUTORY", "MARITAL_STATUS_MARRIED", "MARITAL_STATUS_ANNULLED", "MARITAL_STATUS_OTHER", "MARITAL_STATUS_DOMESTIC_PARTNER", "MARITAL_STATUS_REGISTERED_DOMESTIC_PARTNER", "MARITAL_STATUS_SINGLE", "MARITAL_STATUS_UNREPORTED", "MARITAL_STATUS_WIDOWED"])
-      return false unless marital_status_validator.valid?(@marital_status)
-      true
-    end
-
-    # Custom attribute writer method checking allowed values (enum).
-    # @param [Object] sex Object to be assigned
-    def sex=(sex)
-      validator = EnumAttributeValidator.new('String', ["SEX_UNKNOWN", "SEX_FEMALE", "SEX_MALE", "SEX_OTHER", "SEX_NON_BINARY", "SEX_PREFER_NOT_TO_DISCLOSE"])
-      unless validator.valid?(sex)
-        fail ArgumentError, "invalid value for \"sex\", must be one of #{validator.allowable_values}."
-      end
-      @sex = sex
-    end
-
-    # Custom attribute writer method checking allowed values (enum).
-    # @param [Object] gender Object to be assigned
-    def gender=(gender)
-      validator = EnumAttributeValidator.new('String', ["GENDER_UNKNOWN", "GENDER_FEMALE", "GENDER_MALE", "GENDER_OTHER", "GENDER_NON_BINARY", "GENDER_TRANSGENDER_MALE", "GENDER_TRANSGENDER_FEMALE", "GENDER_PREFER_NOT_TO_DISCLOSE"])
-      unless validator.valid?(gender)
-        fail ArgumentError, "invalid value for \"gender\", must be one of #{validator.allowable_values}."
-      end
-      @gender = gender
-    end
-
-    # Custom attribute writer method checking allowed values (enum).
-    # @param [Object] sexual_orientation Object to be assigned
-    def sexual_orientation=(sexual_orientation)
-      validator = EnumAttributeValidator.new('String', ["ORIENTATION_UNKNOWN", "ORIENTATION_HOMOSEXUAL", "ORIENTATION_HETEROSEXUAL", "ORIENTATION_BISEXUAL", "ORIENTATION_QUESTIONING", "ORIENTATION_OTHER", "ORIENTATION_PREFER_NOT_TO_DISCLOSE"])
-      unless validator.valid?(sexual_orientation)
-        fail ArgumentError, "invalid value for \"sexual_orientation\", must be one of #{validator.allowable_values}."
-      end
-      @sexual_orientation = sexual_orientation
-    end
-
-    # Custom attribute writer method checking allowed values (enum).
-    # @param [Object] ethnicity Object to be assigned
-    def ethnicity=(ethnicity)
-      validator = EnumAttributeValidator.new('String', ["ETHNICITY_UNKNOWN", "ETHNICITY_HISPANIC", "ETHNICITY_NON_HISPANIC", "ETHNICITY_PREFER_NOT_TO_DISCLOSE"])
-      unless validator.valid?(ethnicity)
-        fail ArgumentError, "invalid value for \"ethnicity\", must be one of #{validator.allowable_values}."
-      end
-      @ethnicity = ethnicity
-    end
-
-    # Custom attribute writer method checking allowed values (enum).
-    # @param [Object] marital_status Object to be assigned
-    def marital_status=(marital_status)
-      validator = EnumAttributeValidator.new('String', ["MARITAL_STATUS_UNKNOWN", "MARITAL_STATUS_SEPARATED", "MARITAL_STATUS_UNMARRIED", "MARITAL_STATUS_COMMON_LAW", "MARITAL_STATUS_DIVORCED", "MARITAL_STATUS_LEGALLY_SEPARATED", "MARITAL_STATUS_LIVING_TOGETHER", "MARITAL_STATUS_INTERLOCUTORY", "MARITAL_STATUS_MARRIED", "MARITAL_STATUS_ANNULLED", "MARITAL_STATUS_OTHER", "MARITAL_STATUS_DOMESTIC_PARTNER", "MARITAL_STATUS_REGISTERED_DOMESTIC_PARTNER", "MARITAL_STATUS_SINGLE", "MARITAL_STATUS_UNREPORTED", "MARITAL_STATUS_WIDOWED"])
-      unless validator.valid?(marital_status)
-        fail ArgumentError, "invalid value for \"marital_status\", must be one of #{validator.allowable_values}."
-      end
-      @marital_status = marital_status
-    end
-
-    # Checks equality by comparing each attribute.
-    # @param [Object] Object to be compared
-    def ==(o)
-      return true if self.equal?(o)
-      self.class == o.class &&
-          name == o.name &&
-          dob == o.dob &&
-          ssn == o.ssn &&
-          sex == o.sex &&
-          gender == o.gender &&
-          sexual_orientation == o.sexual_orientation &&
-          races == o.races &&
-          race_details == o.race_details &&
-          ethnicity == o.ethnicity &&
-          ethnicity_detail == o.ethnicity_detail &&
-          marital_status == o.marital_status &&
-          is_deceased == o.is_deceased &&
-          death_date_time == o.death_date_time &&
-          phone_numbers == o.phone_numbers &&
-          email_addresses == o.email_addresses &&
-          language_code == o.language_code &&
-          citizenships == o.citizenships &&
-          address == o.address
-    end
-
-    # @see the `==` method
-    # @param [Object] Object to be compared
-    def eql?(o)
-      self == o
-    end
-
-    # Calculates hash code according to all attributes.
-    # @return [Integer] Hash code
-    def hash
-      [name, dob, ssn, sex, gender, sexual_orientation, races, race_details, ethnicity, ethnicity_detail, marital_status, is_deceased, death_date_time, phone_numbers, email_addresses, language_code, citizenships, address].hash
-    end
-
-    # Builds the object from hash
-    # @param [Hash] attributes Model attributes in the form of hash
-    # @return [Object] Returns the model itself
-    def self.build_from_hash(attributes)
-      new.build_from_hash(attributes)
-    end
-
-    # Builds the object from hash
-    # @param [Hash] attributes Model attributes in the form of hash
-    # @return [Object] Returns the model itself
-    def build_from_hash(attributes)
-      return nil unless attributes.is_a?(Hash)
-      self.class.openapi_types.each_pair do |key, type|
-        if attributes[self.class.attribute_map[key]].nil? && self.class.openapi_nullable.include?(key)
-          self.send("#{key}=", nil)
-        elsif type =~ /\AArray<(.*)>/i
-          # check to ensure the input is an array given that the attribute
-          # is documented as an array but the input is not
-          if attributes[self.class.attribute_map[key]].is_a?(Array)
-            self.send("#{key}=", attributes[self.class.attribute_map[key]].map { |v| _deserialize($1, v) })
-          end
-        elsif !attributes[self.class.attribute_map[key]].nil?
-          self.send("#{key}=", _deserialize(type, attributes[self.class.attribute_map[key]]))
-        end
-      end
-
-      self
-    end
-
-    # Deserializes the data based on type
-    # @param string type Data type
-    # @param string value Value to be deserialized
-    # @return [Object] Deserialized data
-    def _deserialize(type, value)
-      case type.to_sym
-      when :Time
-        Time.parse(value)
-      when :Date
-        Date.parse(value)
-      when :String
-        value.to_s
-      when :Integer
-        value.to_i
-      when :Float
-        value.to_f
-      when :Boolean
-        if value.to_s =~ /\A(true|t|yes|y|1)\z/i
-          true
-        else
-          false
-        end
-      when :Object
-        # generic object (usually a Hash), return directly
-        value
-      when /\AArray<(?<inner_type>.+)>\z/
-        inner_type = Regexp.last_match[:inner_type]
-        value.map { |v| _deserialize(inner_type, v) }
-      when /\AHash<(?<k_type>.+?), (?<v_type>.+)>\z/
-        k_type = Regexp.last_match[:k_type]
-        v_type = Regexp.last_match[:v_type]
-        {}.tap do |hash|
-          value.each do |k, v|
-            hash[_deserialize(k_type, k)] = _deserialize(v_type, v)
-          end
-        end
-      else # model
-        # models (e.g. Pet) or oneOf
-        klass = PrimaryConnectClient.const_get(type)
-        klass.respond_to?(:openapi_one_of) ? klass.build(value) : klass.build_from_hash(value)
-      end
-    end
-
-    # Returns the string representation of the object
-    # @return [String] String presentation of the object
-    def to_s
-      to_hash.to_s
-    end
-
-    # to_body is an alias to to_hash (backward compatibility)
-    # @return [Hash] Returns the object in the form of hash
-    def to_body
-      to_hash
-    end
-
-    # Returns the object in the form of hash
-    # @return [Hash] Returns the object in the form of hash
-    def to_hash
-      hash = {}
-      self.class.attribute_map.each_pair do |attr, param|
-        value = self.send(attr)
-        if value.nil?
-          is_nullable = self.class.openapi_nullable.include?(attr)
-          next if !is_nullable || (is_nullable && !instance_variable_defined?(:"@#{attr}"))
-        end
-
-        hash[param] = _to_hash(value)
-      end
-      hash
-    end
-
-    # Outputs non-array value in the form of hash
-    # For object, use to_hash. Otherwise, just return the value
-    # @param [Object] value Any valid value
-    # @return [Hash] Returns the value in the form of hash
-    def _to_hash(value)
-      if value.is_a?(Array)
-        value.compact.map { |v| _to_hash(v) }
-      elsif value.is_a?(Hash)
-        {}.tap do |hash|
-          value.each { |k, v| hash[k] = _to_hash(v) }
-        end
-      elsif value.respond_to? :to_hash
-        value.to_hash
-      else
-        value
-      end
-    end
-
-  end
-
-end
