+++ lib/primary_connect_client/version.rb
@@ -11,5 +11,5 @@ OpenAPI Generator version: 5.3.0
 =end
 
 module PrimaryConnectClient
-  VERSION = '1.5.0'
+  VERSION = '1.5.1'
 end
