+++ spec/models/subject_spec.rb
@@ -18,14 +18,23 @@ require 'date'
 # Automatically generated by openapi-generator (https://openapi-generator.tech)
 # Please update as you see appropriate
 describe PrimaryConnectClient::Subject do
-  describe '.openapi_one_of' do
-    it 'lists the items referenced in the oneOf array' do
-      expect(described_class.openapi_one_of).to_not be_empty
+  let(:instance) { PrimaryConnectClient::Subject.new }
+
+  describe 'test an instance of Subject' do
+    it 'should create an instance of Subject' do
+      expect(instance).to be_instance_of(PrimaryConnectClient::Subject)
+    end
+  end
+  describe 'test attribute "patient"' do
+    it 'should work' do
+      # assertion here. ref: https://www.relishapp.com/rspec/rspec-expectations/docs/built-in-matchers
     end
   end
 
-  describe '.build' do
-    it 'returns the correct model' do
+  describe 'test attribute "device"' do
+    it 'should work' do
+      # assertion here. ref: https://www.relishapp.com/rspec/rspec-expectations/docs/built-in-matchers
     end
   end
+
 end
